% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{curveRfreq-package}
\alias{curveRfreq}
\alias{curveRfreq-package}
\title{curveRfreq: Frequentist Immunoassay Standard Curve Fitting}
\description{
provides frequentist methods for modeling and calibrating immunoassay standard curves (e.g., 4PL and 5PL models).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
