% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyigompertz4}
\alias{dxdyigompertz4}
\title{\eqn{dx/dy} for the Gompertz Inverse}
\usage{
dxdyigompertz4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Gompertz model parameters.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the Gompertz Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyilogistic4}()},
\code{\link{dxdyilogistic5}()},
\code{\link{dxdyiloglogistic4}()},
\code{\link{dxdyiloglogistic5}()}
}
\concept{dxdy-helpers}
\keyword{internal}
