% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyilogistic5}
\alias{dxdyilogistic5}
\title{\eqn{dx/dy} for the 5PL Inverse}
\usage{
dxdyilogistic5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. 5PL model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the 5PL Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyigompertz4}()},
\code{\link{dxdyilogistic4}()},
\code{\link{dxdyiloglogistic4}()},
\code{\link{dxdyiloglogistic5}()}
}
\concept{dxdy-helpers}
\keyword{internal}
