% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyiloglogistic4}
\alias{dxdyiloglogistic4}
\title{\eqn{dx/dy} for the loglogistic4 Inverse}
\usage{
dxdyiloglogistic4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. 4PL model parameters.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the loglogistic4 Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyigompertz4}()},
\code{\link{dxdyilogistic4}()},
\code{\link{dxdyilogistic5}()},
\code{\link{dxdyiloglogistic5}()}
}
\concept{dxdy-helpers}
\keyword{internal}
