% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyiloglogistic5}
\alias{dxdyiloglogistic5}
\title{\eqn{dx/dy} for the loglogistic5 Inverse}
\usage{
dxdyiloglogistic5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. loglogistic5 model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the loglogistic5 Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyigompertz4}()},
\code{\link{dxdyilogistic4}()},
\code{\link{dxdyilogistic5}()},
\code{\link{dxdyiloglogistic4}()}
}
\concept{dxdy-helpers}
\keyword{internal}
