% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyiY5}
\alias{dxdyiY5}
\title{\eqn{dx/dy} for the 5PL Inverse}
\usage{
dxdyiY5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. 5PL model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the 5PL Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyiY4}()},
\code{\link{dxdyiYd4}()},
\code{\link{dxdyiYd5}()},
\code{\link{dxdyiYgomp4}()}
}
\concept{dxdy-helpers}
\keyword{internal}
