% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyiYd4}
\alias{dxdyiYd4}
\title{\eqn{dx/dy} for the Yd4 Inverse}
\usage{
dxdyiYd4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. 4PL model parameters.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the Yd4 Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyiY4}()},
\code{\link{dxdyiY5}()},
\code{\link{dxdyiYd5}()},
\code{\link{dxdyiYgomp4}()}
}
\concept{dxdy-helpers}
\keyword{internal}
