% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyiYd5}
\alias{dxdyiYd5}
\title{\eqn{dx/dy} for the Yd5 Inverse}
\usage{
dxdyiYd5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Yd5 model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the Yd5 Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyiY4}()},
\code{\link{dxdyiY5}()},
\code{\link{dxdyiYd4}()},
\code{\link{dxdyiYgomp4}()}
}
\concept{dxdy-helpers}
\keyword{internal}
