% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{dxdyiYgomp4}
\alias{dxdyiYgomp4}
\title{\eqn{dx/dy} for the Gompertz Inverse}
\usage{
dxdyiYgomp4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Gompertz model parameters.}
}
\value{
Numeric scalar \eqn{dx/dy}.
}
\description{
\eqn{dx/dy} for the Gompertz Inverse
}
\seealso{
Other dxdy-helpers: 
\code{\link{dxdyiY4}()},
\code{\link{dxdyiY5}()},
\code{\link{dxdyiYd4}()},
\code{\link{dxdyiYd5}()}
}
\concept{dxdy-helpers}
\keyword{internal}
