% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_fitting.R
\name{get_blank_se}
\alias{get_blank_se}
\title{Extract the Standard Error of the Blank from Antigen Settings}
\usage{
get_blank_se(antigen_settings)
}
\arguments{
\item{antigen_settings}{Named list (output of
\code{\link{obtain_lower_constraint}}).}
}
\value{
Numeric scalar: \code{std_error_blank}.
}
\description{
Extract the Standard Error of the Blank from Antigen Settings
}
