% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{grad_inv_loglogistic5_fixed}
\alias{grad_inv_loglogistic5_fixed}
\title{Gradient of the Inverse loglogistic5 with Fixed Lower Asymptote}
\usage{
grad_inv_loglogistic5_fixed(y, fixed_a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{fixed_a}{Numeric scalar. Externally fixed lower asymptote.}

\item{b, c, d}{Numeric scalars. Free model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter (free).}
}
\value{
Named numeric vector \code{c(b=, c=, d=, g=)}.
}
\description{
Gradient of the Inverse loglogistic5 with Fixed Lower Asymptote
}
\seealso{
Other fixed-a-gradients: 
\code{\link{grad_inv_gompertz4_fixed}()},
\code{\link{grad_inv_logistic4_fixed}()},
\code{\link{grad_inv_logistic5_fixed}()},
\code{\link{grad_inv_loglogistic4_fixed}()}
}
\concept{fixed-a-gradients}
\keyword{internal}
