% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{grad_loglogistic4}
\alias{grad_loglogistic4}
\title{Analytical Gradient of the Inverse loglogistic4}
\usage{
grad_loglogistic4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free loglogistic4 model parameters.}
}
\value{
A list with \code{grad_theta} (named vector \code{c(a=, b=, c=, d=)})
and scalar \code{grad_y}.
}
\description{
Returns partial derivatives for \code{\link[=inv_loglogistic4]{inv_loglogistic4()}} with all parameters free.
}
\seealso{
\code{\link[=inv_loglogistic4]{inv_loglogistic4()}}, \code{\link[=grad_inv_loglogistic4_fixed]{grad_inv_loglogistic4_fixed()}}

Other gradient-functions: 
\code{\link{grad_gompertz4}()},
\code{\link{grad_logistic4}()},
\code{\link{grad_logistic5}()},
\code{\link{grad_loglogistic5}()}
}
\concept{gradient-functions}
