% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{grad_loglogistic5}
\alias{grad_loglogistic5}
\title{Analytical Gradient of the Inverse loglogistic5}
\usage{
grad_loglogistic5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free loglogistic5 model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
A list with \code{grad_theta} (named vector \code{c(a=, b=, c=, d=, g=)})
and scalar \code{grad_y}.
}
\description{
Returns partial derivatives for \code{\link[=inv_loglogistic5]{inv_loglogistic5()}} with all five parameters free.
}
\seealso{
\code{\link[=inv_loglogistic5]{inv_loglogistic5()}}, \code{\link[=grad_inv_loglogistic5_fixed]{grad_inv_loglogistic5_fixed()}}

Other gradient-functions: 
\code{\link{grad_gompertz4}()},
\code{\link{grad_logistic4}()},
\code{\link{grad_logistic5}()},
\code{\link{grad_loglogistic4}()}
}
\concept{gradient-functions}
