% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{grad_y_Yd4_fixed}
\alias{grad_y_Yd4_fixed}
\title{\eqn{\partial x / \partial y} for the Yd4 Inverse with Fixed \eqn{a}}
\usage{
grad_y_Yd4_fixed(y, fixed_a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{fixed_a}{Numeric scalar. Externally fixed lower asymptote.}

\item{b, c, d}{Numeric scalars. Free model parameters.}
}
\value{
Numeric scalar \eqn{\partial x / \partial y}.
}
\description{
\eqn{\partial x / \partial y} for the Yd4 Inverse with Fixed \eqn{a}
}
\seealso{
Other fixed-a-grad-y: 
\code{\link{grad_y_Y4_fixed}()},
\code{\link{grad_y_Y5_fixed}()},
\code{\link{grad_y_Yd5_fixed}()},
\code{\link{grad_y_Ygomp4_fixed}()}
}
\concept{fixed-a-grad-y}
\keyword{internal}
