% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{grad_Yd4}
\alias{grad_Yd4}
\title{Analytical Gradient of the Inverse Yd4}
\usage{
grad_Yd4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free Yd4 model parameters.}
}
\value{
A list with \code{grad_theta} (named vector \code{c(a=, b=, c=, d=)})
and scalar \code{grad_y}.
}
\description{
Returns partial derivatives for \code{\link[=inv_Yd4]{inv_Yd4()}} with all parameters free.
}
\seealso{
\code{\link[=inv_Yd4]{inv_Yd4()}}, \code{\link[=grad_inv_Yd4_fixed]{grad_inv_Yd4_fixed()}}

Other gradient-functions: 
\code{\link{grad_Y4}()},
\code{\link{grad_Y5}()},
\code{\link{grad_Yd5}()},
\code{\link{grad_Ygomp4}()}
}
\concept{gradient-functions}
