% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{grad_Ygomp4}
\alias{grad_Ygomp4}
\title{Analytical Gradient of the Inverse Gompertz}
\usage{
grad_Ygomp4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric scalar. Observed response.}

\item{a, b, c, d}{Numeric scalars. Free Gompertz model parameters.}
}
\value{
A list with \code{grad_theta} (named vector \code{c(a=, b=, c=, d=)})
and scalar \code{grad_y}.
}
\description{
Returns partial derivatives for \code{\link[=inv_Ygomp4]{inv_Ygomp4()}} with all parameters free.
}
\seealso{
\code{\link[=inv_Ygomp4]{inv_Ygomp4()}}, \code{\link[=grad_inv_Ygomp4_fixed]{grad_inv_Ygomp4_fixed()}}

Other gradient-functions: 
\code{\link{grad_Y4}()},
\code{\link{grad_Y5}()},
\code{\link{grad_Yd4}()},
\code{\link{grad_Yd5}()}
}
\concept{gradient-functions}
