% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_gompertz4}
\alias{inv_gompertz4}
\title{Inverse of the Gompertz Model}
\usage{
inv_gompertz4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. Gompertz model parameters.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under \code{\link[=gompertz4]{gompertz4()}}:
\deqn{x = c - \frac{1}{b}\,\log\!\left(-\log\frac{y - a}{d - a}\right)}
}
\seealso{
\code{\link[=gompertz4]{gompertz4()}}, \code{\link[=inv_gompertz4_fixed]{inv_gompertz4_fixed()}}

Other inverse-functions: 
\code{\link{inv_gompertz4_fixed}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_logistic4_fixed}()},
\code{\link{inv_logistic5}()},
\code{\link{inv_logistic5_fixed}()},
\code{\link{inv_loglogistic4}()},
\code{\link{inv_loglogistic4_fixed}()},
\code{\link{inv_loglogistic5}()},
\code{\link{inv_loglogistic5_fixed}()}
}
\concept{inverse-functions}
