% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_logistic5}
\alias{inv_logistic5}
\title{Inverse of the 5PL Model}
\usage{
inv_logistic5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. 5PL model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under \code{\link[=logistic5]{logistic5()}}:
\deqn{x = c + b\,\log\!\left(\left(\frac{a - d}{y - d}\right)^{1/g}
  - 1\right)}
}
\seealso{
\code{\link[=logistic5]{logistic5()}}, \code{\link[=inv_logistic5_fixed]{inv_logistic5_fixed()}}

Other inverse-functions: 
\code{\link{inv_gompertz4}()},
\code{\link{inv_gompertz4_fixed}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_logistic4_fixed}()},
\code{\link{inv_logistic5_fixed}()},
\code{\link{inv_loglogistic4}()},
\code{\link{inv_loglogistic4_fixed}()},
\code{\link{inv_loglogistic5}()},
\code{\link{inv_loglogistic5_fixed}()}
}
\concept{inverse-functions}
