% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_loglogistic4}
\alias{inv_loglogistic4}
\title{Inverse of the loglogistic4 Model}
\usage{
inv_loglogistic4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. loglogistic4 model parameters.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under the \code{\link[=loglogistic4]{loglogistic4()}} parameterisation:
\deqn{x = c\,\left(\frac{d - a}{y - a} - 1\right)^{1/b}}
}
\seealso{
\code{\link[=loglogistic4]{loglogistic4()}}, \code{\link[=inv_loglogistic4_fixed]{inv_loglogistic4_fixed()}}

Other inverse-functions: 
\code{\link{inv_gompertz4}()},
\code{\link{inv_gompertz4_fixed}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_logistic4_fixed}()},
\code{\link{inv_logistic5}()},
\code{\link{inv_logistic5_fixed}()},
\code{\link{inv_loglogistic4_fixed}()},
\code{\link{inv_loglogistic5}()},
\code{\link{inv_loglogistic5_fixed}()}
}
\concept{inverse-functions}
