% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_loglogistic5}
\alias{inv_loglogistic5}
\title{Inverse of the loglogistic5 Model}
\usage{
inv_loglogistic5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. loglogistic5 model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under \code{\link[=loglogistic5]{loglogistic5()}}:
\deqn{x = c - \frac{1}{b}\left(\log\!\left(\left(\frac{y - a}{d - a}
  \right)^{-g} - 1\right) - \log g\right)}
}
\seealso{
\code{\link[=loglogistic5]{loglogistic5()}}, \code{\link[=inv_loglogistic5_fixed]{inv_loglogistic5_fixed()}}

Other inverse-functions: 
\code{\link{inv_gompertz4}()},
\code{\link{inv_gompertz4_fixed}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_logistic4_fixed}()},
\code{\link{inv_logistic5}()},
\code{\link{inv_logistic5_fixed}()},
\code{\link{inv_loglogistic4}()},
\code{\link{inv_loglogistic4_fixed}()},
\code{\link{inv_loglogistic5_fixed}()}
}
\concept{inverse-functions}
