% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_loglogistic5_fixed}
\alias{inv_loglogistic5_fixed}
\title{Inverse of the loglogistic5 Model with Fixed Lower Asymptote}
\usage{
inv_loglogistic5_fixed(y, fixed_a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{fixed_a}{Numeric scalar. Externally fixed lower asymptote.}

\item{b, c, d}{Numeric scalars. Model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Inverse of the loglogistic5 Model with Fixed Lower Asymptote
}
\seealso{
Other inverse-functions: 
\code{\link{inv_gompertz4}()},
\code{\link{inv_gompertz4_fixed}()},
\code{\link{inv_logistic4}()},
\code{\link{inv_logistic4_fixed}()},
\code{\link{inv_logistic5}()},
\code{\link{inv_logistic5_fixed}()},
\code{\link{inv_loglogistic4}()},
\code{\link{inv_loglogistic4_fixed}()},
\code{\link{inv_loglogistic5}()}
}
\concept{inverse-functions}
\keyword{internal}
