% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_Y5}
\alias{inv_Y5}
\title{Inverse of the 5PL Model}
\usage{
inv_Y5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. 5PL model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under \code{\link[=Y5]{Y5()}}:
\deqn{x = c + b\,\log\!\left(\left(\frac{a - d}{y - d}\right)^{1/g}
  - 1\right)}
}
\seealso{
\code{\link[=Y5]{Y5()}}, \code{\link[=inv_Y5_fixed]{inv_Y5_fixed()}}

Other inverse-functions: 
\code{\link{inv_Y4}()},
\code{\link{inv_Y4_fixed}()},
\code{\link{inv_Y5_fixed}()},
\code{\link{inv_Yd4}()},
\code{\link{inv_Yd4_fixed}()},
\code{\link{inv_Yd5}()},
\code{\link{inv_Yd5_fixed}()},
\code{\link{inv_Ygomp4}()},
\code{\link{inv_Ygomp4_fixed}()}
}
\concept{inverse-functions}
