% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_Yd4}
\alias{inv_Yd4}
\title{Inverse of the Yd4 Model}
\usage{
inv_Yd4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. Yd4 model parameters.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under the \code{\link[=Yd4]{Yd4()}} parameterisation:
\deqn{x = c\,\left(\frac{d - a}{y - a} - 1\right)^{1/b}}
}
\seealso{
\code{\link[=Yd4]{Yd4()}}, \code{\link[=inv_Yd4_fixed]{inv_Yd4_fixed()}}

Other inverse-functions: 
\code{\link{inv_Y4}()},
\code{\link{inv_Y4_fixed}()},
\code{\link{inv_Y5}()},
\code{\link{inv_Y5_fixed}()},
\code{\link{inv_Yd4_fixed}()},
\code{\link{inv_Yd5}()},
\code{\link{inv_Yd5_fixed}()},
\code{\link{inv_Ygomp4}()},
\code{\link{inv_Ygomp4_fixed}()}
}
\concept{inverse-functions}
