% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_Yd5}
\alias{inv_Yd5}
\title{Inverse of the Yd5 Model}
\usage{
inv_Yd5(y, a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. Yd5 model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under \code{\link[=Yd5]{Yd5()}}:
\deqn{x = c - \frac{1}{b}\left(\log\!\left(\left(\frac{y - a}{d - a}
  \right)^{-g} - 1\right) - \log g\right)}
}
\seealso{
\code{\link[=Yd5]{Yd5()}}, \code{\link[=inv_Yd5_fixed]{inv_Yd5_fixed()}}

Other inverse-functions: 
\code{\link{inv_Y4}()},
\code{\link{inv_Y4_fixed}()},
\code{\link{inv_Y5}()},
\code{\link{inv_Y5_fixed}()},
\code{\link{inv_Yd4}()},
\code{\link{inv_Yd4_fixed}()},
\code{\link{inv_Yd5_fixed}()},
\code{\link{inv_Ygomp4}()},
\code{\link{inv_Ygomp4_fixed}()}
}
\concept{inverse-functions}
