% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_Yd5_fixed}
\alias{inv_Yd5_fixed}
\title{Inverse of the Yd5 Model with Fixed Lower Asymptote}
\usage{
inv_Yd5_fixed(y, fixed_a, b, c, d, g)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{fixed_a}{Numeric scalar. Externally fixed lower asymptote.}

\item{b, c, d}{Numeric scalars. Model parameters.}

\item{g}{Numeric scalar. Asymmetry parameter.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Inverse of the Yd5 Model with Fixed Lower Asymptote
}
\seealso{
Other inverse-functions: 
\code{\link{inv_Y4}()},
\code{\link{inv_Y4_fixed}()},
\code{\link{inv_Y5}()},
\code{\link{inv_Y5_fixed}()},
\code{\link{inv_Yd4}()},
\code{\link{inv_Yd4_fixed}()},
\code{\link{inv_Yd5}()},
\code{\link{inv_Ygomp4}()},
\code{\link{inv_Ygomp4_fixed}()}
}
\concept{inverse-functions}
\keyword{internal}
