% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{inv_Ygomp4}
\alias{inv_Ygomp4}
\title{Inverse of the Gompertz Model}
\usage{
inv_Ygomp4(y, a, b, c, d)
}
\arguments{
\item{y}{Numeric vector. Observed response values.}

\item{a, b, c, d}{Numeric scalars. Gompertz model parameters.}
}
\value{
Numeric vector of estimated \code{x} values.
}
\description{
Solves for \code{x} given response \code{y} under \code{\link[=Ygomp4]{Ygomp4()}}:
\deqn{x = c - \frac{1}{b}\,\log\!\left(-\log\frac{y - a}{d - a}\right)}
}
\seealso{
\code{\link[=Ygomp4]{Ygomp4()}}, \code{\link[=inv_Ygomp4_fixed]{inv_Ygomp4_fixed()}}

Other inverse-functions: 
\code{\link{inv_Y4}()},
\code{\link{inv_Y4_fixed}()},
\code{\link{inv_Y5}()},
\code{\link{inv_Y5_fixed}()},
\code{\link{inv_Yd4}()},
\code{\link{inv_Yd4_fixed}()},
\code{\link{inv_Yd5}()},
\code{\link{inv_Yd5_fixed}()},
\code{\link{inv_Ygomp4_fixed}()}
}
\concept{inverse-functions}
