% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_wavelength}
\alias{normalize_wavelength}
\title{Normalize wavelength value for bead-arrays.}
\usage{
normalize_wavelength(x)
}
\arguments{
\item{x}{Character or NA. Wavelength value to normalize.}
}
\value{
Character vector.
}
\description{
Replace NA or empty wavelength values with the sentinel.
}
