% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value.}

\item{b}{Fallback value if \code{a} is \code{NULL}.}
}
\value{
The first non-NULL value.
}
\description{
Returns the left-hand side if it is not \code{NULL}, otherwise returns the right-hand side.
}
\examples{
NULL \%||\% 5
10 \%||\% 5

}
