% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{WL_NONE}
\alias{WL_NONE}
\title{Sentinel value for missing wavelength}
\format{
A character scalar.
}
\usage{
WL_NONE
}
\description{
Constant used to represent a missing or undefined wavelength in grouping keys.
This avoids the use of \code{NA} or \code{NULL}, ensuring compatibility with SQL
\code{UNIQUE} constraints and stable joins in R workflows, particularly for
bead-array data where wavelength may not be defined.
}
\details{
This value must match the sentinel used for assays that do
not have a wavelength (bead arrays)
}
\examples{
WL_NONE

}
\keyword{datasets}
